requirements.txt
python-dotenv==1.0.1
boto3==1.34.82
apache-airflow-providers-apache-spark==4.1.5
pyspark==3.5.1